requirements.txt
langchain==0.0.255
streamlit==1.22.0
newsapi-python
langchain-groq